// src/types/info.ts
export interface PersonalInfo {
  firstName: string;
  lastName: string;
  email: string;
  phoneNumber: string;
}

export interface EducationalInfo {
  schoolName: string;
  titleOfStudy: string;
  dateStudy: string;
}

export interface PracticalInfo {
  companyName: string;
  positionTitle: string;
  mainResponsabilities: string;
  startDateExp: string;
  endDateExp: string;
}